Add `TrieCommitment` trait and `VerkleCommitment` type
This introduces a new `TrieCommitment` trait for the generic handling of
commitment updates in managed tries. Implementors of the trait receive
notifications about children/slots being changed, which are the two
conditions in which a commitment will have to be recomputed.

For the Verkle trie implementation, a new `VerkleCommitment` type is
introduced which implements the trait. It keeps track of the
commitment's dirty state, as well as the used slots (required for leaf
commitments) and which children/slots have changed since the last
commitment computation.