Fix unfortunate bug in leaf commitment computation
THIS FIX COSTS ~40 MGas/s :((

This moves the leaf commitment computation into a separate function
(compute_leaf_node_commitment_v2) and fixes a very unfortunate bug:
If c was not the default commitment (= the common case), it would be
updated using the delta of c1/prev_c1 and c2/prev_c2.
When not using batch updates, c1/c2 are updated in place.
The bug: We copied prev_1 / prev_2 AFTER doing the in-place update,
thereby making the delta 0.